requirements.txt
Django # psycopg, .env loading, jinja2 not included
uvicorn
Jinja2
pydantic # serialization
loguru
python-dotenv # no need for uvicorn deployment
psycopg[binary,pool] # prebuilt, pool compatible